fix(chat): address Copilot + Greptile review comments on #749
Three reviewer findings addressed:

1. DiscoTimeoutError reported the module-level DISCO_IQ_TIMEOUT_MS in
   its message, but `withIqTimeout` may have used an explicit override
   (e.g. 30ms in tests via __setDiscoIqTimeoutForTest). The error
   message would still read "timed out after 15000ms" — misleading at
   the diagnostic point. Now the constructor takes the effective
   timeout and stores it on the error, so logs always reflect the
   actual budget the timer fired against.

2. `hydrateRoomInfo`'s internal try/catch made the allSettled rejection
   branches in `discoverChannels` and `discoverTopology` unreachable
   dead code. Removed the inner catch — `Promise.allSettled` at the
   caller is now the single resilience layer, with the unhydrated
   `channelFromRoom` fallback fired by genuine rejection paths
   (e.g. a timed-out sendDiscoInfo).

3. FakeDOMParser + the XML fixture helpers were duplicated between
   `discovery-pin-permission.test.ts` and `discovery-resilience.test.ts`
   (~110 lines each). Extracted to `tests/helpers/disco-xml.ts` and
   imported from both — no more drift surface.

bun test: 1117 pass / 1 pre-existing fail. bun run lint: clean.
