+++ chat/tests/discovery-resilience.test.ts
@@ -5,6 +5,12 @@ import {
   discoverTopology,
   withIqTimeout,
 } from "../src/lib/xmpp/discovery";
+import {
+  discoInfoXml,
+  discoItemsXml,
+  pubsubItemsXml,
+  withFakeDomParser,
+} from "./helpers/disco-xml";
 
 afterEach(() => {
   // Belt-and-braces: every test that mutates the disco timeout restores it
@@ -301,114 +307,3 @@ function resilientClient(options: ResilientClientOptions = {}) {
   };
 }
 
-function discoItemsXml(items: Array<{ name?: string; node?: string; jid?: string }>): string {
-  return `<iq type="result"><query xmlns="http://jabber.org/protocol/disco#items">${items.map((item) =>
-    `<item${item.jid ? ` jid="${item.jid}"` : ""}${item.name ? ` name="${item.name}"` : ""}${item.node ? ` node="${item.node}"` : ""}/>`
-  ).join("")}</query></iq>`;
-}
-
-function discoInfoXml(info: { features?: string[]; identities?: Array<{ category: string; type: string; name?: string }> } = {}): string {
-  return `<iq type="result"><query xmlns="http://jabber.org/protocol/disco#info">${
-    (info.identities ?? []).map((identity) =>
-      `<identity category="${identity.category}" type="${identity.type}"${identity.name ? ` name="${identity.name}"` : ""}/>`
-    ).join("")
-  }${
-    (info.features ?? []).map((feature) => `<feature var="${feature}"/>`).join("")
-  }</query></iq>`;
-}
-
-function pubsubItemsXml(items: Array<{ id?: string; jid?: string; name?: string }>): string {
-  return `<iq type="result"><pubsub xmlns="http://jabber.org/protocol/pubsub"><items>${items.map((item) =>
-    `<item${item.id ? ` id="${item.id}"` : ""}>${item.jid || item.name ? `<conference${item.jid ? ` jid="${item.jid}"` : ""}${item.name ? ` name="${item.name}"` : ""}/>` : ""}</item>`
-  ).join("")}</items></pubsub></iq>`;
-}
-
-async function withFakeDomParser(run: () => Promise<void>): Promise<void> {
-  const original = globalThis.DOMParser;
-  globalThis.DOMParser = FakeDOMParser as typeof DOMParser;
-  try {
-    await run();
-  } finally {
-    globalThis.DOMParser = original;
-  }
-}
-
-class FakeDOMParser {
-  parseFromString(xml: string): Document {
-    return parseTestXml(xml) as unknown as Document;
-  }
-}
-
-class FakeXmlElement {
-  readonly children: FakeXmlElement[] = [];
-  text = "";
-  parentNode: FakeXmlElement | null = null;
-
-  constructor(
-    readonly localName: string,
-    readonly namespaceURI: string | null,
-    private readonly attrs: Record<string, string>,
-  ) {}
-
-  get textContent(): string {
-    return `${this.text}${this.children.map((child) => child.textContent ?? "").join("")}`;
-  }
-
-  getAttribute(name: string): string | null {
-    return this.attrs[name] ?? null;
-  }
-
-  querySelector(localName: string): FakeXmlElement | null {
-    return this.getElementsByTagName(localName)[0] ?? null;
-  }
-
-  getElementsByTagNameNS(namespace: string, localName: string): FakeXmlElement[] {
-    return this.descendants().filter((node) => node.namespaceURI === namespace && node.localName === localName);
-  }
-
-  getElementsByTagName(localName: string): FakeXmlElement[] {
-    return this.descendants().filter((node) => node.localName === localName);
-  }
-
-  private descendants(): FakeXmlElement[] {
-    return this.children.flatMap((child) => [child, ...child.descendants()]);
-  }
-}
-
-function parseTestXml(xml: string): FakeXmlElement {
-  const root = new FakeXmlElement("#document", null, {});
-  const stack: FakeXmlElement[] = [root];
-  const tokenPattern = /<[^>]+>|[^<]+/g;
-  for (const token of xml.match(tokenPattern) ?? []) {
-    if (token.startsWith("</")) {
-      stack.pop();
-      continue;
-    }
-    if (token.startsWith("<")) {
-      const selfClosing = token.endsWith("/>");
-      const body = token.slice(1, selfClosing ? -2 : -1).trim();
-      const [qualifiedName = ""] = body.split(/\s+/, 1);
-      const attrs = attributesFromTag(body);
-      const parent = stack[stack.length - 1];
-      const node = new FakeXmlElement(
-        qualifiedName.includes(":") ? qualifiedName.split(":").pop()! : qualifiedName,
-        attrs.xmlns ?? parent.namespaceURI,
-        attrs,
-      );
-      node.parentNode = parent;
-      parent.children.push(node);
-      if (!selfClosing) stack.push(node);
-      continue;
-    }
-    stack[stack.length - 1].text += token;
-  }
-  return root;
-}
-
-function attributesFromTag(tagBody: string): Record<string, string> {
-  const attrs: Record<string, string> = {};
-  for (const match of tagBody.matchAll(/([A-Za-z0-9:_-]+)="([^"]*)"/g)) {
-    attrs[match[1]] = match[2];
-  }
-  return attrs;
-}
